requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
transformers
torch
huggingface_hub 
python-dotenv
# accelerate
# bitsandbytes
# torchvision
google-generativeai
google
mcp
mcp-server
sentence-transformers
faiss-cpu
# torch==2.1.0+cpu
# torchvision==0.16.0+cpu
langsmith